Product price base type

+++ htdocs/install/mysql/tables/llx_product_attribute_combination.sql
@@ -18,11 +18,11 @@
 
 CREATE TABLE llx_product_attribute_combination
 (
-  rowid INT PRIMARY KEY NOT NULL AUTO_INCREMENT,
-  fk_product_parent INT NOT NULL,
-  fk_product_child INT NOT NULL,
-  variation_price FLOAT NOT NULL,
-  variation_price_percentage INT NULL,
-  variation_weight FLOAT NOT NULL,
-  entity INT DEFAULT 1 NOT NULL
+  rowid INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+  fk_product_parent INTEGER NOT NULL,
+  fk_product_child INTEGER NOT NULL,
+  variation_price DOUBLE(24,8) NOT NULL,
+  variation_price_percentage INTEGER NULL,
+  variation_weight REAL NOT NULL,
+  entity INTEGER DEFAULT 1 NOT NULL
 )ENGINE=innodb;
